pyproject.toml
[project]
name = "telbot-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-minio-backend>=3.8.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
]
